Modularize transcript compression
